refactor: move all the lib types into one file
They will be generated in the future.
