+++ src/lib/types.ts
@@ -25,3 +25,47 @@ export interface Favorite {
 export type SettingsPayload = { [key: string]: string };
 
 // TODO: add more types from lib/types
+
+export interface ActiveJob {
+  object_id: string;
+  status: string;
+}
+
+export interface ObjectIcon {
+  icon: {
+    emoji: string;
+    type: string;
+    external: {
+      url: string;
+    };
+    file: {
+      url: string;
+    };
+  };
+}
+
+export interface ObjectTitle {
+  title: [
+    {
+      type: string;
+      text: { content: string; href: string };
+      plain_text: string;
+    }
+  ];
+  object: string;
+  properties: {
+    Name: { title: { plain_text: string } };
+    title: {
+      title: [{ type: string; text: { content: string }; plain_text: string }];
+    };
+  };
+}
+
+export interface TemplateFile {
+  parent: string;
+  name: string;
+  front: string;
+  back: string;
+  styling: string;
+  storageKey: string;
+}
