--- src/lib/types/ActiveJob.ts
@@ -1,6 +1,0 @@
-interface ActiveJob {
-  object_id: string;
-  status: string;
-}
-
-export default ActiveJob;
