--- src/lib/types/ObjectIcon.ts
@@ -1,14 +1,0 @@
-interface ObjectIcon {
-  icon: {
-    emoji: string;
-    type: string;
-    external: {
-      url: string;
-    };
-    file: {
-      url: string;
-    };
-  };
-}
-
-export default ObjectIcon;
