--- src/lib/types/TemplateFile.ts
@@ -1,10 +1,0 @@
-interface TemplateFile {
-  parent: string;
-  name: string;
-  front: string;
-  back: string;
-  styling: string;
-  storageKey: string;
-}
-
-export default TemplateFile;
